fix: set poll status to true (#16)
the pollMap status check implemented in PR #6 was only set to true
during the first run and not any subsequent runs. this fixes that so
that it is always set to true when proceeding with the poll.

move the defer to after the the "still running" return, otherwise a
second poll after a long running poll would cause the pollMap to get set
to false when the initial thread may still be running.